+++ src/arch/Makefile.mk
@@ -8,8 +8,8 @@ abt_sources += \
 	arch/abtd_affinity_parser.c \
 	arch/abtd_env.c \
 	arch/abtd_stream.c \
-	arch/abtd_thread.c \
-	arch/abtd_time.c
+	arch/abtd_time.c \
+	arch/abtd_ythread.c
 
 if ABT_USE_FCONTEXT
 abt_sources += \
