+++ src/include/abtd.h
@@ -76,7 +76,7 @@ void ABTD_affinity_list_free(ABTD_affinity_list *p_list);
 #include "abtd_stream.h"
 
 /* ULT Context */
-#include "abtd_thread.h"
+#include "abtd_ythread.h"
 void ABTD_ythread_exit(ABTI_xstream *p_local_xstream, ABTI_ythread *p_ythread);
 void ABTD_ythread_cancel(ABTI_xstream *p_local_xstream,
                          ABTI_ythread *p_ythread);
