+++ src/include/abti.h
@@ -623,11 +623,11 @@ void ABTI_info_check_print_all_thread_stacks(void);
 #include "abti_pool.h"
 #include "abti_sched.h"
 #include "abti_config.h"
-#include "abti_unit.h"
 #include "abti_stream.h"
 #include "abti_self.h"
-#include "abti_tool.h"
 #include "abti_thread.h"
+#include "abti_tool.h"
+#include "abti_ythread.h"
 #include "abti_thread_attr.h"
 #include "abti_task.h"
 #include "abti_mutex.h"
